refactor: Create `ProductList` component to be used by `ProductLanding` page component
